call save_pages_to_db from one way
