+++ core/utils/sqlite.py
@@ -16,3 +16,16 @@ def create_database_table(db_name):
         '''CREATE TABLE IF NOT EXISTS pages (id INTEGER PRIMARY KEY AUTOINCREMENT, title TEXT, status INTEGER, date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,thread INTEGER)''')
 
     return conn, cursor
+
+
+def save_pages_to_db(gen, conn, cursor, thread_number):
+    for entry in gen:
+        try:
+            title = entry
+            cursor.execute("SELECT * FROM pages WHERE title = ?", (title,))
+            if cursor.fetchone() is None:
+                print("added : " + title)
+                cursor.execute("INSERT INTO pages (title, status,thread) VALUES (?, 0,?)", (title, int(thread_number)))
+            conn.commit()
+        except sqlite3.Error as e:
+            print(f"An error occurred while inserting the title {entry.title()} into the database: {e}")
