+++ tasks/maintenance/module.py
@@ -62,19 +62,6 @@ FROM (
     return gen
 
 
-def save_pages_to_db(gen, conn, cursor, thread_number):
-    for entry in gen:
-        try:
-            title = entry
-            cursor.execute("SELECT * FROM pages WHERE title = ?", (title,))
-            if cursor.fetchone() is None:
-                print("added : " + title)
-                cursor.execute("INSERT INTO pages (title, status,thread) VALUES (?, 0,?)", (title, int(thread_number)))
-            conn.commit()
-        except sqlite3.Error as e:
-            print(f"An error occurred while inserting the title {entry.title()} into the database: {e}")
-
-
 def get_articles(cursor, thread_number):
     random_number = random.randint(1, 10)
     cursor.execute("SELECT id, title,thread FROM pages WHERE thread=? and status=0 ORDER BY date ASC LIMIT 50 OFFSET ?;", (int(thread_number),int(random_number)))
