+++ tasks/webcite/module.py
@@ -34,19 +34,6 @@ FROM (
     return gen
 
 
-def save_pages_to_db(gen, conn, cursor, thread_number):
-    for entry in gen:
-        try:
-            title = entry
-            cursor.execute("SELECT * FROM pages WHERE title = ?", (title,))
-            if cursor.fetchone() is None:
-                print("added : " + title)
-                cursor.execute("INSERT INTO pages (title, status,thread) VALUES (?, 0,?)",
-                               (title, int(thread_number)))
-            conn.commit()
-        except sqlite3.Error as e:
-            print(f"An error occurred while inserting the title {entry.title()} into the database: {e}")
-
 
 def get_articles(cursor, thread_number):
     if thread_number == 1:
